Expose selections from code host location.hash to extensions
