+++ shared/src/util/url.test.ts
@@ -1,5 +1,6 @@
 import {
     buildSearchURLQuery,
+    lprToSelectionsZeroIndexed,
     makeRepoURI,
     parseHash,
     parseRepoURI,
@@ -205,6 +206,8 @@ describe('util/url', () => {
         test('parses hash with range', () => {
             expect(parseHash('L1-2')).toEqual({ line: 1, endLine: 2 })
             expect(parseHash('L1:2-3:4')).toEqual({ line: 1, character: 2, endLine: 3, endCharacter: 4 })
+            expect(parseHash('L47-L55')).toEqual({ line: 47, endLine: 55 })
+            expect(parseHash('L34:2-L38:3')).toEqual({ line: 34, character: 2, endLine: 38, endCharacter: 3 })
         })
 
         test('parses hash with references', () => {
@@ -278,4 +281,127 @@ describe('buildSearchURLQuery', () => {
     it('handles characters that need encoding', () =>
         expect(buildSearchURLQuery('foo bar%baz')).toBe('q=foo+bar%25baz'))
     it('preserves / and : for readability', () => expect(buildSearchURLQuery('repo:foo/bar')).toBe('q=repo:foo/bar'))
+})
+
+describe('lprToSelectionsZeroIndexed', () => {
+    test('converts an LPR with only a start line', () => {
+        assertDeepStrictEqual(
+            lprToSelectionsZeroIndexed({
+                line: 5,
+            }),
+            [
+                {
+                    start: {
+                        line: 4,
+                        character: 0,
+                    },
+                    end: {
+                        line: 4,
+                        character: 0,
+                    },
+                    anchor: {
+                        line: 4,
+                        character: 0,
+                    },
+                    active: {
+                        line: 4,
+                        character: 0,
+                    },
+                    isReversed: false,
+                },
+            ]
+        )
+    })
+
+    test('converts an LPR with a line and a character', () => {
+        assertDeepStrictEqual(
+            lprToSelectionsZeroIndexed({
+                line: 5,
+                character: 45,
+            }),
+            [
+                {
+                    start: {
+                        line: 4,
+                        character: 44,
+                    },
+                    end: {
+                        line: 4,
+                        character: 44,
+                    },
+                    anchor: {
+                        line: 4,
+                        character: 44,
+                    },
+                    active: {
+                        line: 4,
+                        character: 44,
+                    },
+                    isReversed: false,
+                },
+            ]
+        )
+    })
+
+    test('converts an LPR with a start and end line', () => {
+        assertDeepStrictEqual(
+            lprToSelectionsZeroIndexed({
+                line: 12,
+                endLine: 15,
+            }),
+            [
+                {
+                    start: {
+                        line: 11,
+                        character: 0,
+                    },
+                    end: {
+                        line: 14,
+                        character: 0,
+                    },
+                    anchor: {
+                        line: 11,
+                        character: 0,
+                    },
+                    active: {
+                        line: 14,
+                        character: 0,
+                    },
+                    isReversed: false,
+                },
+            ]
+        )
+    })
+
+    test('converts an LPR with a start and end line and characters', () => {
+        assertDeepStrictEqual(
+            lprToSelectionsZeroIndexed({
+                line: 12,
+                character: 30,
+                endLine: 15,
+                endCharacter: 60,
+            }),
+            [
+                {
+                    start: {
+                        line: 11,
+                        character: 29,
+                    },
+                    end: {
+                        line: 14,
+                        character: 59,
+                    },
+                    anchor: {
+                        line: 11,
+                        character: 29,
+                    },
+                    active: {
+                        line: 14,
+                        character: 59,
+                    },
+                    isReversed: false,
+                },
+            ]
+        )
+    })
 })
