+++ web/src/repo/blob/actions/ToggleDiscussions.tsx
@@ -4,10 +4,14 @@ import { fromEvent, Subject, Subscription } from 'rxjs'
 import { filter } from 'rxjs/operators'
 import { ChatIcon } from '../../../../../shared/src/components/icons' // TODO: Switch to mdi icon
 import { LinkOrButton } from '../../../../../shared/src/components/LinkOrButton'
-import { parseHash, toPositionOrRangeHash, toViewStateHashComponent } from '../../../../../shared/src/util/url'
+import {
+    lprToRange,
+    parseHash,
+    toPositionOrRangeHash,
+    toViewStateHashComponent,
+} from '../../../../../shared/src/util/url'
 import { Tooltip } from '../../../components/tooltip/Tooltip'
 import { eventLogger } from '../../../tracking/eventLogger'
-import { lprToRange } from '../../../util/url'
 import { BlobPanelTabID } from '../panel/BlobPanel'
 /**
  * A repository header action that toggles the visibility of the discussions panel.
